fix bug

+++ attention/common/op_kernel/arch32/fia_block_cube_nonquant.h
@@ -49,7 +49,7 @@ public:
     static constexpr bool ANTIQUANT = !IsSameType<Q_T, KV_T>::value;
     static constexpr bool KVINT4 = IsSameType<KV_T, int4b_t>::value;
     static constexpr bool QUANT = (IsSameType<Q_T, KV_T>::value && IsSameType<KV_T, int8_t>::value);
-    static constexpr uint32_t BLOCK_SIZE_ALIGN_16 = 16U;
+    static constexpr uint64_t BLOCK_SIZE_ALIGN_16 = 16UL;
     // define pse datetype
     using pseShiftType = typename AscendC::Conditional<AscendC::IsSameType<Q_T, int8_t>::value, half, Q_T>::type;
     // 后接量化的条件需要重新审视
